l7rrt封装和测试
**现象及复现步骤**

无

**原因和解决方案**

**影响范围**

无

**验证方案**

无

**涉及分支**

* master

**检查项**

- [ ] 需要同步修改(trident-dump、trident-copy或trident-testsrv)
- [ ] 需要更新依赖
- [ ] 是共性问题(代码中存在类似问题)
- [ ] 打包成功(如果有修改Makefile)

/assign @jinjie

/cc @zhichong @yuanchao @jinjie

+++ src/flow_generator/flow_map.rs
@@ -900,7 +900,7 @@ impl FlowMap {
             flow.flow_perf_stats = node.meta_flow_perf.as_mut().and_then(|perf| {
                 perf.copy_and_reset_perf_data(
                     flow.reversed,
-                    l7_timeout_count,
+                    l7_timeout_count as u32,
                     self.l4_metrics_enabled(),
                     self.l7_metrics_enabled(),
                 )
